JFrame => JDialog, setLocation() => setLocationRelativeTo()
Replaced some JFrames by JDialogs (which are more appropriate in this
case)

Added modality for JDialogs, makes underlying frames unusable while the
dialogs are shown.

Replaced ALL setLocation() by setLocationRelativeTo(), which removes the
necessity of calculations : this method already handles them. Plus the
JDialogs now pop where they are supposed to and not just in the middle
of the screen. (their location are relative to their parent for
CreditFrame and OptionFrame)